Move some top-level code under shared
