+++ benchmarks/postgres/helpers/piscina.ts
@@ -1,7 +1,7 @@
 import Piscina from '@posthog/piscina'
 import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 
-import { defaultConfig } from '../../../src/config'
+import { defaultConfig } from '../../../src/shared/config'
 import { UUIDT } from '../../../src/shared/utils'
 import { LogLevel } from '../../../src/types'
 import { makePiscina } from '../../../src/worker/piscina'
