+++ benchmarks/vm/memory.benchmark.ts
@@ -5,7 +5,7 @@ import { Plugin, PluginConfig, PluginConfigVMReponse } from '../../src/types'
 import { createPluginConfigVM } from '../../src/worker/vm/vm'
 import { commonOrganizationId } from '../../tests/helpers/plugins'
 
-jest.mock('../../src/sql')
+jest.mock('../../src/shared/sql')
 jest.setTimeout(600000) // 600 sec timeout
 
 function createEvent(index: number): PluginEvent {
