+++ benchmarks/vm/worker.benchmark.ts
@@ -2,12 +2,12 @@ import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 import * as os from 'os'
 import { performance } from 'perf_hooks'
 
-import { defaultConfig } from '../../src/config'
+import { defaultConfig } from '../../src/shared/config'
 import { LogLevel } from '../../src/types'
 import { makePiscina } from '../../src/worker/piscina'
 import { resetTestDatabase } from '../../tests/helpers/sql'
 
-jest.mock('../../src/sql')
+jest.mock('../../src/shared/sql')
 jest.setTimeout(600000) // 600 sec timeout
 
 function processOneEvent(
