+++ src/server.ts
@@ -14,13 +14,13 @@ import * as path from 'path'
 import { Pool, types as pgTypes } from 'pg'
 import { ConnectionOptions } from 'tls'
 
-import { defaultConfig } from './config'
-import { DB } from './db'
 import { startQueue } from './main/queue'
 import { startSchedule } from './main/services/schedule'
 import { startFastifyInstance, stopFastifyInstance } from './main/web/server'
+import { defaultConfig } from './shared/config'
+import { DB } from './shared/db'
+import { status } from './shared/status'
 import { createPostgresPool, createRedis, delay, UUIDT } from './shared/utils'
-import { status } from './status'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from './types'
 import { EventsProcessor } from './worker/ingestion/process-event'
