+++ src/shared/db.ts
@@ -7,16 +7,6 @@ import { Producer, ProducerRecord } from 'kafkajs'
 import { DateTime } from 'luxon'
 import { Pool, PoolClient, QueryConfig, QueryResult, QueryResultRow } from 'pg'
 
-import { KAFKA_PERSON, KAFKA_PERSON_UNIQUE_ID } from './shared/ingestion/topics'
-import { chainToElements, hashElements, timeoutGuard, unparsePersonPartial } from './shared/ingestion/utils'
-import {
-    castTimestampOrNow,
-    clickHouseTimestampToISO,
-    createRedis,
-    escapeClickHouseString,
-    sanitizeSqlIdentifier,
-    tryTwice,
-} from './shared/utils'
 import {
     ClickHouseEvent,
     ClickHousePerson,
@@ -33,7 +23,17 @@ import {
     RawPerson,
     SessionRecordingEvent,
     TimestampFormat,
-} from './types'
+} from '../types'
+import { KAFKA_PERSON, KAFKA_PERSON_UNIQUE_ID } from './ingestion/topics'
+import { chainToElements, hashElements, timeoutGuard, unparsePersonPartial } from './ingestion/utils'
+import {
+    castTimestampOrNow,
+    clickHouseTimestampToISO,
+    createRedis,
+    escapeClickHouseString,
+    sanitizeSqlIdentifier,
+    tryTwice,
+} from './utils'
 
 /** The recommended way of accessing the database. */
 export class DB {
