+++ tests/helpers/sql.ts
@@ -1,6 +1,6 @@
 import { Pool, PoolClient } from 'pg'
 
-import { defaultConfig } from '../../src/config'
+import { defaultConfig } from '../../src/shared/config'
 import { delay, UUIDT } from '../../src/shared/utils'
 import { PluginsServer, PluginsServerConfig, Team } from '../../src/types'
 import { commonOrganizationId, commonOrganizationMembershipId, commonUserId, makePluginObjects } from './plugins'
