+++ tests/schedule.test.ts
@@ -14,8 +14,8 @@ import { createPromise } from './helpers/promises'
 import { resetTestDatabase } from './helpers/sql'
 import { setupPiscina } from './helpers/worker'
 
-jest.mock('../src/sql')
-jest.mock('../src/status')
+jest.mock('../src/shared/sql')
+jest.mock('../src/shared/status')
 jest.setTimeout(60000) // 60 sec timeout
 
 function createEvent(index = 0): PluginEvent {
