+++ tests/server.test.ts
@@ -5,7 +5,7 @@ import { LogLevel } from '../src/types'
 import { makePiscina } from '../src/worker/piscina'
 import { resetTestDatabase } from './helpers/sql'
 
-jest.mock('../src/sql')
+jest.mock('../src/shared/sql')
 jest.setTimeout(60000) // 60 sec timeout
 
 test('startPluginsServer', async () => {
